2018.05.29.17.49: done. Real environment validation

+++ AutoChangeConfig/AutoChangeConfig.py
@@ -1,3 +1,4 @@
+# !/usr/local/python
 # -*- coding: UTF-8 -*-
 import os
 import shutil
@@ -12,7 +13,10 @@ from xml.etree import ElementTree
 class AutoChangeConfig(object):
     def __init__(self):
         self.modifypath = ''
-        self.notes_list = []
+        self.xml_notes_list = []
+        self.other_notes_list = []
+        self.ini_notes_list = []
+        self.ini_notes_tmp_list = []
         # Changing the local path to the current text path
         os.chdir(os.path.split(os.path.realpath(__file__))[0])
         self.home = os.getcwd()
@@ -26,7 +30,7 @@ class AutoChangeConfig(object):
                   (logfloder, self.time_now)
         logging.basicConfig(filename=logfile, level=logging.DEBUG,
                             format='%(asctime)s %(name)s %(lineno)d %(levelname)s %(message)s',
-                            datefmt='%Y-%m-%d-%a %H:%M:%S', filemode='w')
+                            datefmt='%Y-%m-%d-%a %H:%M:%S', filemode='wb')
         self.loggerchangeconfig = logging.getLogger('ChangeConfig')
         formatter = logging.Formatter('%(asctime)s %(name)s %(lineno)d %(levelname)s  %(message)s')
         logscr = logging.StreamHandler()
@@ -66,10 +70,14 @@ class AutoChangeConfig(object):
                 self.lookupannotations(path, suffix)
                 self.modifyxml(path, changed)
             elif 'properties' == suffix:
+                self.lookupannotations(path, suffix)
                 self.modifyproperties(path, changed)
             elif 'ini' == suffix:
+                self.lookupannotations(path, suffix)
+                self.ini_notes_tmp_list = self.ini_notes_list
                 self.modifyini(path, changed, modify_node)
             elif 'cfg' == suffix:
+                self.lookupannotations(path, suffix)
                 self.modifycfg(path, changed)
             else:
                 self.loggerchangeconfig.error('There is no suffix matching, Please check the configuration file')
@@ -91,11 +99,12 @@ class AutoChangeConfig(object):
             node_str = '[' + str(modify_node.find('nodename').text) + ']'
             self.loggerchangeconfig.info('Begin modify ini file')
             self.loggerchangeconfig.info('The modify ini file is \"%s\"' % path)
-            with open(path, 'r') as f:
+            with open(path, 'rb') as f:
                 cfg_content = f.read()
             while True:
                 n += 1
-                with open(path, 'r') as fr:
+                self.lookupannotations(path, 'ini')
+                with open(path, 'rb') as fr:
                     cfg_content2 = fr.read()
                 if str(cfg_content2)[begin_num_list_tmp[-1]:].find(node_str) >= 0:
                     begin_num = int(str(cfg_content2)[begin_num_list_tmp[-1]:].find(node_str)) + len(node_str) + \
@@ -104,15 +113,19 @@ class AutoChangeConfig(object):
                     if -1 == str(cfg_content2[begin_num:]).find('['):
                         end_num = len(cfg_content2)
                     else:
-                        end_num = str(cfg_content2[begin_num:]).find('[') + begin_num
+                        end_num = str(cfg_content2[begin_num:]).find('[') + begin_num - 1
                     # print 'begin_num is %d' % begin_num
                     # print 'end_num: %d' % end_first_num
                     # print '----'
                     # print begin_num_list
                     # print '++++'
                     # print end_num_list
-                    with open(path, 'w') as fw:
-                        fw.write(cfg_content2.replace(cfg_content2[begin_num:end_num], '\nautochangeconfig%d\n' % n))
+                    with open(path, 'wb') as fw:
+                        if not 'false' == self.judgerange(begin_num, self.ini_notes_list):
+                            fw.write(cfg_content2.replace(cfg_content2[begin_num:end_num],
+                                                          '\nautochangeconfig%d\n' % n))
+                        else:
+                            fw.write(cfg_content2)
 
                 else:
                     break
@@ -141,19 +154,21 @@ class AutoChangeConfig(object):
 
             del(begin_num_list[0])
             for i in range(0, len(begin_num_list)):
-                with open(path, 'r') as fr:
-                    cfg_content3 = fr.read()
-                tmpfile_path = os.path.join(self.home, 'temp')
-                with open(tmpfile_path, 'w') as fw:
-                    # print str(cfg_content[begin_num_list[i]:end_num_list[i]])
-                    fw.write(str(cfg_content[begin_num_list[i]:end_num_list[i]]))
-                self.modifyotherfile(tmpfile_path, changed)
-                with open(tmpfile_path, 'r') as fr:
-                    fr_str = fr.read()
-                # print fr_str
-                with open(path, 'w') as fw:
-                    fw.write(str(cfg_content3).replace('autochangeconfig%d' % (i + 1), fr_str))
-                os.remove(tmpfile_path)
+                if not 'false' == self.judgerange(begin_num_list[i], self.ini_notes_tmp_list):
+                    with open(path, 'rb') as fr:
+                        cfg_content3 = fr.read()
+                    tmpfile_path = os.path.join(self.home, 'temp')
+                    with open(tmpfile_path, 'wb') as fw:
+                        # print str(cfg_content[begin_num_list[i]:end_num_list[i]])
+                        fw.write(str(cfg_content[begin_num_list[i]:end_num_list[i]-1]))
+                    self.lookupannotations(tmpfile_path, 'other')
+                    self.modifyotherfile(tmpfile_path, changed)
+                    with open(tmpfile_path, 'rb') as fr:
+                        fr_str = fr.read()
+                    # print fr_str
+                    with open(path, 'wb') as fw:
+                        fw.write(str(cfg_content3).replace('\nautochangeconfig%d\n' % (i + 1), fr_str))
+                    os.remove(tmpfile_path)
             self.loggerchangeconfig.info('modify ini file \"%s\" done' % path)
         else:
             self.loggerchangeconfig.info('Begin modify ini file')
@@ -172,7 +187,7 @@ class AutoChangeConfig(object):
         self.loggerchangeconfig.info('The modify xml file is \"%s\"' % path)
         for change_str in changed:
             # Read the original text by line and get the list of text.
-            with open(path, 'r') as fr:
+            with open(path, 'rb') as fr:
                 b_list = fr.readlines()
             changed_list = []
             original_list = []
@@ -193,13 +208,15 @@ class AutoChangeConfig(object):
             for l in range(0, len(original_list)):
                 if str(original_list[l]) == str(changed_list[0]):
                     # Lists all rows that modify the contents of the first row in the original article.
-                    first_line_num_list.append(l)
+                    if not 'false' == self.judgerange(l, self.xml_notes_list):
+                        first_line_num_list.append(l)
             changed_list_lines = len(changed_list)
             for i in range(0, changed_list_lines):
                 try:
                     # List all the lines that modify the content in the original article.
                     previous_line_num += original_list[previous_line_num:].index(changed_list[i])
-                    number_list.append(previous_line_num)
+                    if not 'false' == self.judgerange(previous_line_num, self.xml_notes_list):
+                        number_list.append(previous_line_num)
                 except Exception, xml_err:
                     if xml_err:
                         pass
@@ -210,20 +227,22 @@ class AutoChangeConfig(object):
                 self.loggerchangeconfig.error('Can\'t find content, please check your configuration')
                 return
 
-            with open(path, 'r') as fr:
+            with open(path, 'rb') as fr:
                 c_list = fr.readlines()
 
-            for m in range(0, len(str(change_str.text).splitlines(True))):
+            for m in range(0, len(str(change_str.text).splitlines(True)) - 1):
                 # Modify the corresponding content in the original text
+                if m == len(str(change_str.text).splitlines(True)) - 1:
+                    str(change_str.text).splitlines(True)[m] += '\n'
                 c_list[last_first_line+m] = str(change_str.text).splitlines(True)[m]
 
-            with open(path, 'w') as fw:
+            with open(path, 'wb') as fw:
                 fw.writelines(c_list)
         self.loggerchangeconfig.info('Modify xml file done.')
 
     def modifyotherfile(self, path, changed):
         cfg_content_new = ''
-        with open(path, 'r') as f:
+        with open(path, 'rb') as f:
             cfg_content = f.readlines()
         for change_str in changed:
             if '' != cfg_content_new:
@@ -232,7 +251,7 @@ class AutoChangeConfig(object):
             cfg_content_new = self.replacestring(change_str.text, cfg_content)
 
         # print '++++' + str(cfg_content_new) + '++++'
-        with open(path, 'w') as f2:
+        with open(path, 'wb') as f2:
             for line in cfg_content_new:
                 f2.write(line)
 
@@ -253,13 +272,14 @@ class AutoChangeConfig(object):
                     self.loggerchangeconfig.info('Now modify %s' % str(str(replacestr).split(symbol)[0]))
                     for i in range(0, len(originalcontent)):
                         # Look for a similar string in each line
-                        begin_num = str(originalcontent[i]).find(str(replacestr).split(symbol)[0] + symbol)
-                        if begin_num >= 0:
-                            # print '>>>' + str(symbol) + '<<<'
-                            # print '>>>' + str(begin_num) + '<<<'
-                            # print '>>>' + str(replacestr) + '<<<'
-                            originalcontent[i] = str(replacestr) + '\n'
-                            find_key = 'true'
+                        if not 'false' == self.judgerange(i, self.other_notes_list):
+                            begin_num = str(originalcontent[i]).find(str(replacestr).split(symbol)[0] + symbol)
+                            if begin_num == 0:
+                                # print '>>>' + str(symbol) + '<<<'
+                                # print '>>>' + str(begin_num) + '<<<'
+                                # print '>>>' + str(replacestr) + '<<<'
+                                originalcontent[i] = str(replacestr) + '\n'
+                                find_key = 'true'
                     if 'false' == find_key:
                         # print '>>>' + str(replacestr) + '<<<'
                         # print originalcontent
@@ -276,8 +296,8 @@ class AutoChangeConfig(object):
     def getlinenumber(self, num_list, first_line_num_list):
         d_value1 = 0
         last_first_num = 0
-        print num_list
-        print first_line_num_list
+        # print num_list
+        # print first_line_num_list
         for num in first_line_num_list:
             if num <= num_list[-1]:
                 d_value2 = int(num_list[-1]) - int(num)
@@ -332,7 +352,7 @@ class AutoChangeConfig(object):
         self.loggerchangeconfig.info('Restore config file done')
 
     def lookupannotations(self, path, suffix):
-        with open(path, 'r') as fr:
+        with open(path, 'rb') as fr:
             notes_str = fr.readlines()
         if 'xml' == suffix:
             c_list = []
@@ -344,8 +364,47 @@ class AutoChangeConfig(object):
                     d_list.append(n)
                     c_list.append(d_list)
                     d_list = []
-            print c_list
-            self.notes_list = c_list
+            # print c_list
+            self.xml_notes_list = c_list
+        elif 'ini' == suffix:
+            with open(path, 'rb') as fr:
+                ini_notes_str = fr.read()
+            c_list = []
+            d_list = []
+            notes_end_num = 0
+            while True:
+                notes_begin_num = ini_notes_str[notes_end_num:].find('#')
+                if notes_begin_num >= 0:
+                    notes_begin_tmp_num = notes_begin_num + notes_end_num
+                    d_list.append(notes_begin_tmp_num)
+                    notes_end_num = ini_notes_str[notes_begin_tmp_num:].find('\n')
+                    notes_end_tmp_num = notes_begin_tmp_num + notes_end_num
+                    notes_end_num = notes_end_tmp_num
+                    d_list.append(notes_end_tmp_num)
+                    c_list.append(d_list)
+                    d_list = []
+                else:
+                    break
+            # print c_list
+            self.ini_notes_list = c_list
+        else:
+            c_list = []
+            d_list = []
+            for n in range(0, len(notes_str)):
+                if str(notes_str[n]).find('#') >= 0:
+                    d_list.append(n)
+                    d_list.append(n)
+                    c_list.append(d_list)
+                    d_list = []
+            # print c_list
+            self.other_notes_list = c_list
+
+    @staticmethod
+    def judgerange(num, lst):
+        for o in range(0, len(lst)):
+            if num >= lst[o][0]:
+                if num <= lst[o][1]:
+                    return 'false'
 
     def main(self):
         try:
